requirements.txt
absl-py==0.11.0
antlr4-python3-runtime==4.8
argon2-cffi==20.1.0
articleDateExtractor==0.20
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.1
bert-serving-client==1.10.0
bert-serving-server==1.10.0
bleach==3.2.1
bleu==0.3
blis==0.4.1
boto3==1.15.14
botocore==1.18.14
brotlipy==0.7.0
bs4==0.0.1
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.3
Cython==0.29.21
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
dython==0.6.1
easydict==1.9
editdistance==0.3.1
efficiency==0.4
empath==0.89
entrypoints==0.3
fairseq==0.9.0
fastBPE==0.1.0
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gensim==3.8.3
google==3.0.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googlesearch-python==1.1.0
GPUtil==1.4.0
grpcio==1.32.0
h5py==2.10.0
hatesonar==0.0.7
hydra-core==1.0.3
idna==2.10
imbalanced-learn==0.7.0
imblearn==0.0
importlib-metadata==2.0.0
importlib-resources==3.3.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab==2.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
language-check==1.1
lxml==4.6.2
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
murmurhash==1.0.2
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
neuralcoref==4.0
nlpnet==1.2.4
nltk==3.5
notebook==6.1.4
numpy==1.19.5
oauthlib==3.1.0
omegaconf==2.0.3
opt-einsum==3.3.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
patsy==0.5.1
pandas==1.2.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
portalocker==2.0.0
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.14.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.20
pyenchant==3.1.1
Pygments==2.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
pytrends==4.7.3
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
regex==2020.9.27
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.6
ruamel-yaml==0.15.87
s3transfer==0.3.3
sacrebleu==1.4.14
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.7
seaborn==0.11.0
Send2Trash==1.5.0
sentencepiece==0.1.91
six==1.15.0
smart-open==3.0.0
soupsieve==2.0.1
spacy==2.3.2
srsly==1.0.2
statsmodels==0.12.2
subword-nmt==0.3.7
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
thinc==7.4.1
threadpoolctl==2.1.0
tldextract==2.2.3
tokenizers==0.8.1rc2
torch==1.7.1
torchtext==0.8.1
tornado==6.0.4
tqdm==4.46.0
traitlets==5.0.4
transformers==3.3.1
tweepy==3.10.0
typing-extensions==3.7.4.3
urllib3==1.25.9
unidecode==1.1.2
vaderSentiment==3.3.2
wasabi==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.0
wrapt==1.12.1
zipp==3.3.0
